requirements.txt
# Core dependencies
asyncio-mqtt==0.16.1
httpx==0.25.2
aiofiles==23.2.1
click==8.1.7
rich==13.7.0
pydantic==2.5.0
pyyaml==6.0.1
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1

# AI/ML dependencies
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
spacy==3.7.2
pytesseract==0.3.10

# Security testing
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Fuzzing and payload generation
faker==20.1.0

# Reporting
weasyprint==60.2
markdown==3.5.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Async task queue
celery==5.3.4
redis==5.0.1